New fixture in pytest-simcore to extract examples
